emmm

+++ bv-api/func.c
@@ -10,15 +10,15 @@ static int64_t xor = INT64_C(177451812);
 static int64_t add = INT64_C(100618342136696320);
 
 void bvInitialize(void) {
-	for (register unsigned int i=0; i<58; i++) {
+	for (register uint_fast16_t i=0; i<58; i++) {
 		tr[(size_t)table[i]] = i;
 	}
 }
 
-inline static int64_t bvi64Pow(int64_t x, unsigned int a) {
-	register unsigned int aa = a;
+inline static int64_t bvi64Pow(int64_t x, uint_fast16_t a) {
+	register uint_fast16_t aa = a;
 	register int64_t s = 1;
-	for (register unsigned int i=0; i<aa; i++) {
+	for (register uint_fast16_t i=0; i<aa; i++) {
 		s *= x;
 	}
 	return s;
@@ -26,7 +26,7 @@ inline static int64_t bvi64Pow(int64_t x, unsigned int a) {
 
 int64_t bvDec(char x[]) {
 	register int64_t r = 0;
-	for (register unsigned int i=0; i<10; i++) {
+	for (register uint_fast16_t i=0; i<10; i++) {
 		r += tr[(size_t)x[s[i]]*bvi64Pow(58,i)];
 	}
 	return (r-add)^xor;
@@ -34,7 +34,7 @@ int64_t bvDec(char x[]) {
 
 char * bvEnc(int64_t x) {
 	x = (x^xor) + add;
-	for (register unsigned int i=0; i<10; i++) {
+	for (register uint_fast16_t i=0; i<10; i++) {
 		r[s[i]] = table[x/bvi64Pow(58,i)%58];
 	}
 	return r;
